Still working on getting the restore functions working 	modified:   src/Python PYSAT_GUI/pysat_main.py
Note: This is what can and can't be used
QSettings is designed for saving menu items and values.
IT DOES NOT SAVE THE ACTUAL EXISTENCE OF THOSE OBJECTS
It simply saves what was in the objects.
The one thing you can use this for is to load data that is inside boxes.
Honestly no different than what you were hoping to accomplish anyways.
Which is a list of functions that load the entire UI in
and then setText()ing everything