Adding progress bar back in for reading ChemCam data
the progress bar object is an optional input to io script, so it is
possible to run the script without it. I moved the PyQt5 import down
into ccam_batch to accomodate this. It's a little awkward, but makes it
possible for the program to not use PyQt5 if desired.