Finished most of the required notebook cells.
All of the cells except for the "How to export the result to a GeoTiff with spatial information" cell are now implemented, along with a few additional cells, as requested by Dr. Gaddis. 
These include: "How to Export Any Kind of Figure to a .png File" and "How to Plot Multiple Spectra in One Figure."

Due to my lack of knowledge regarding some of the required cells, in retrospect, I don't think the implementation is complete for
"How to Extract Spectra with a Function Applied" and "How to apply the Horgan algorithms."
With the former, I am not sure how to extract spectra *after* a function is applied, since almost all of the functions (including band depth) convert the given 3-D array to a 2-D array.
With the latter, I learned that I am only generating the parameters of the Horgan Algorithm, rather than encompassing the entire process. As such, I have renamed this cell to: "How to Obtain Horgan Algorithm Parameters."

Otherwise, I feel that this notebook is very close to completion.